requirements.txt
crewai>=0.1.0
'crewai[tools]'
exa-py>=1.0.0
firecrawl-py>=1.9.0
composio-phidata>=0.1.0
streamlit>=1.24.0
python-dotenv>=0.19.0